store/requirements.txt
acme==2.11.0
asgiref==3.8.1
Babel==2.15.0
backports.zoneinfo==0.2.1
certbot==2.11.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ConfigArgParse==1.7
configobj==5.0.8
cryptography==43.0.0
distro==1.9.0
Django==4.2
django-extra-views==0.14.0
django-haystack==3.3.0
django-oscar==3.2.4
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
factory-boy==3.2.1
Faker==26.0.0
gunicorn==22.0.0
html5lib==1.1
idna==3.7
importlib_metadata==8.1.0
importlib_resources==6.4.0
josepy==1.14.0
packaging==24.1
parsedatetime==2.6
phonenumbers==8.13.40
pillow==10.4.0
pkg_resources==0.0.0
psycopg2-binary==2.9.9
purl==1.6
pycountry==24.6.1
pycparser==2.22
pyOpenSSL==24.2.1
pyRFC3339==1.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
zipp==3.19.2
